backend/requirements.txt
argcomplete==3.6.3
blinker==1.9.0
boto3==1.42.58
botocore==1.42.58
certifi==2026.2.25
cfn-flip==1.3.0
charset-normalizer==3.4.4
click==8.3.1
durationpy==0.10
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
hjson==3.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
packaging==26.0
placebo==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
setuptools==79.0.1
six==1.17.0
SQLAlchemy==2.0.47
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.3
troposphere==4.10.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
zappa==0.61.4
